REPO-3811 Refactor Spring context into Java beans
