+++ src/main/java/org/alfresco/repo/events/raw/Event.java
@@ -23,35 +23,22 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.repo.rawevents.types;
+package org.alfresco.repo.events.raw;
 
 /**
- * Event that is raised once per node when any of the content properties on the
- * node are changed; the specific properties are irrelevant. This is primarily
- * useful to determine when a new file is introduced into the system.
+ * This interface defines the basic data which is captured for each event.
  */
-public class OnContentUpdatePolicyEvent extends EventImpl
+public interface Event
 {
-    private String nodeRef;
-    private boolean newContent;
+    String getId();
 
-    public String getNodeRef()
-    {
-        return nodeRef;
-    }
+    String getType();
 
-    public void setNodeRef(String nodeRef)
-    {
-        this.nodeRef = nodeRef;
-    }
+    String getAuthenticatedUser();
 
-    public boolean isNewContent()
-    {
-        return newContent;
-    }
+    String getExecutingUser();
 
-    public void setNewContent(boolean newContent)
-    {
-        this.newContent = newContent;
-    }
+    Long getTimestamp();
+
+    int getSchema();
 }
