+++ src/main/java/org/alfresco/repo/events/raw/OnContentUpdatePolicyEvent.java
@@ -23,14 +23,35 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.repo.rawevents.types;
+package org.alfresco.repo.events.raw;
 
-import java.io.Serializable;
-
-public enum EventType implements Serializable
+/**
+ * Event that is raised once per node when any of the content properties on the
+ * node are changed; the specific properties are irrelevant. This is primarily
+ * useful to determine when a new file is introduced into the system.
+ */
+public class OnContentUpdatePolicyEvent extends EventImpl
 {
-    /**
-     * Event type that should be used when creating a content update event.
-     */
-    CONTENT_UPDATED
+    private String nodeRef;
+    private boolean newContent;
+
+    public String getNodeRef()
+    {
+        return nodeRef;
+    }
+
+    public void setNodeRef(String nodeRef)
+    {
+        this.nodeRef = nodeRef;
+    }
+
+    public boolean isNewContent()
+    {
+        return newContent;
+    }
+
+    public void setNewContent(boolean newContent)
+    {
+        this.newContent = newContent;
+    }
 }
