+++ src/main/java/org/alfresco/repo/messaging/camel/producer/TransactionAwareEventProducer.java
@@ -23,7 +23,7 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.repo.rawevents;
+package org.alfresco.repo.messaging.camel.producer;
 
 import java.util.LinkedList;
 import java.util.List;
