+++ src/main/resources/alfresco/application-context-highlevel.xml
@@ -42,7 +42,5 @@
     <import resource="classpath:alfresco/heartbeat/heartbeat-repo-collector-context.xml"/>
     <import resource="classpath*:alfresco/enterprise/heartbeat/heartbeat-enterprise-repo-collector-context.xml" />
     <import resource="classpath*:alfresco/deprecated-model-bootstrap-context.xml" />
-    <import resource="classpath*:alfresco/events-context.xml" />
     <import resource="classpath*:alfresco/messaging-context.xml" />
-    <import resource="classpath*:alfresco/desktop-sync-messaging-context.xml" />
 </beans>
