+++ src/main/resources/alfresco/events-context.xml
@@ -2,9 +2,9 @@
 <!DOCTYPE beans PUBLIC '-//SPRING//DTD BEAN//EN' 'http://www.springframework.org/dtd/spring-beans.dtd'>
 
 <beans>
-    <bean id="Events" class="org.alfresco.repo.management.subsystems.ChildApplicationContextFactory" parent="abstractPropertyBackedBean">
-        <property name="autoStart">
-            <value>${events.subsystem.autoStart}</value>
-        </property>
-    </bean>
+    <!--<bean id="Events" class="org.alfresco.repo.management.subsystems.ChildApplicationContextFactory" parent="abstractPropertyBackedBean">-->
+        <!--<property name="autoStart">-->
+            <!--<value>${events.subsystem.autoStart}</value>-->
+        <!--</property>-->
+    <!--</bean>-->
 </beans>
