+++ src/main/resources/alfresco/messaging-context.xml
@@ -2,96 +2,17 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:camel="http://camel.apache.org/schema/spring"
        xsi:schemaLocation="
         http://www.springframework.org/schema/beans
         http://www.springframework.org/schema/beans/spring-beans.xsd
         http://www.springframework.org/schema/context
-        http://www.springframework.org/schema/context/spring-context.xsd
-        http://camel.apache.org/schema/spring
-        http://camel.apache.org/schema/spring/camel-spring.xsd">
+        http://www.springframework.org/schema/context/spring-context.xsd">
 
-    <context:component-scan base-package="org.alfresco.messaging.camel.configuration"/>
-
-    <bean id="defaultDataFormat" class="org.apache.camel.component.jackson.JacksonDataFormat">
-        <constructor-arg ref="alfrescoEventObjectMapper" />
-        <constructor-arg value="java.lang.Object" />
-    </bean>
-
-    <bean id="pooledConnectionFactory" class="org.apache.activemq.pool.PooledConnectionFactory"
-          init-method="start" destroy-method="stop">
-        <property name="maxConnections" value="${messaging.broker.connections.max}" />
-        <property name="connectionFactory" ref="activeMqConnectionFactory" />
-    </bean>
-
-    <bean id="messagingTransactionManager" class="org.springframework.jms.connection.JmsTransactionManager">
-        <property name="connectionFactory" ref="pooledConnectionFactory"/>
-    </bean>
-
-    <bean id="amqp" class="org.apache.camel.component.amqp.AMQPComponent">
-        <property name="connectionFactory" ref="pooledConnectionFactory" />
-        <property name="transacted" value="${messaging.transacted}" />
-        <property name="transactionManager" ref="messagingTransactionManager" />
-    </bean>
-
-    <bean id="CAMEL_PROPAGATION_MANDATORY" class="org.apache.camel.spring.spi.SpringTransactionPolicy">
-        <property name="transactionManager" ref="messagingTransactionManager" />
-        <property name="propagationBehaviorName" value="PROPAGATION_MANDATORY" />
-    </bean>
-
-    <bean id="CAMEL_PROPAGATION_NEVER" class="org.apache.camel.spring.spi.SpringTransactionPolicy">
-        <property name="transactionManager" ref="messagingTransactionManager" />
-        <property name="propagationBehaviorName" value="PROPAGATION_NEVER" />
-    </bean>
-
-    <bean id="CAMEL_PROPAGATION_NOT_SUPPORTED" class="org.apache.camel.spring.spi.SpringTransactionPolicy">
-        <property name="transactionManager" ref="messagingTransactionManager" />
-        <property name="propagationBehaviorName" value="PROPAGATION_NOT_SUPPORTED" />
-    </bean>
-
-    <bean id="CAMEL_PROPAGATION_REQUIRED" class="org.apache.camel.spring.spi.SpringTransactionPolicy">
-        <property name="transactionManager" ref="messagingTransactionManager" />
-        <property name="propagationBehaviorName" value="PROPAGATION_REQUIRED" />
-    </bean>
-
-    <bean id="CAMEL_PROPAGATION_REQUIRES_NEW" class="org.apache.camel.spring.spi.SpringTransactionPolicy">
-        <property name="transactionManager" ref="messagingTransactionManager" />
-        <property name="propagationBehaviorName" value="PROPAGATION_REQUIRES_NEW" />
-    </bean>
-
-    <bean id="CAMEL_PROPAGATION_SUPPORTS" class="org.apache.camel.spring.spi.SpringTransactionPolicy">
-        <property name="transactionManager" ref="messagingTransactionManager" />
-        <property name="propagationBehaviorName" value="PROPAGATION_SUPPORTS" />
-    </bean>
-
-    <!-- Any RouteBuilder in this package will be fired up -->
-    <context:component-scan base-package="org.alfresco.messaging.camel.routes"/>
-
-    <!-- Any RouteBuilder in this enterprise package will be fired up -->
-    <!--<context:component-scan base-package="org.alfresco.enterprise.repo.routes"/>-->
-
-    <!--<camelContext id="alfrescoCamelContext" xmlns="http://camel.apache.org/schema/spring">-->
-        <!--<contextScan/>-->
-        <!--<camel:jmxAgent id="agent" mbeanObjectDomainName="Alfresco.Camel" />-->
-        <!--<template id="camelProducerTemplate" defaultEndpoint="direct:alfresco.default" />-->
-    <!--</camelContext>-->
-
-    <!--<import resource="defaultCamelRoutes.xml" />-->
-    <!--<bean id="defaultRouteLoader" class="org.alfresco.messaging.camel.SpringContextRouteLoader">-->
-        <!--<property name="camelContextId" value="alfrescoCamelContext" />-->
-        <!--<property name="routeContextId" value="deadLetterRoutes" />-->
-    <!--</bean>-->
-
-    <context:component-scan base-package="org.alfresco.repo.rawevents"/>
-
-    <!--<bean id="alfrescoEventObjectMapper" class="org.springframework.http.converter.json.Jackson2ObjectMapperFactoryBean">-->
-        <!--<property name="failOnEmptyBeans" value="false"/>-->
-        <!--<property name="indentOutput" value="true"/>-->
-    <!--</bean>-->
-
-    <!--<bean id="transactionAwareEventProducer" class="org.alfresco.repo.rawevents.TransactionAwareEventProducer">-->
-        <!--<property name="producer" ref="producer" />-->
-        <!--<property name="endpoint" value="direct-vm:alfresco.raw.events" />-->
-        <!--<property name="objectMapper" ref="objectMapper"/>-->
-    <!--</bean>-->
+    <!-- Base Camel configuration -->
+    <context:component-scan base-package="org.alfresco.repo.messaging.camel.configuration"/>
+    <!-- Camel Producer -->
+    <context:component-scan base-package="org.alfresco.repo.messaging.camel.producer"/>
+    <!-- Camel Routes definitions -->
+    <context:component-scan base-package="org.alfresco.repo.messaging.camel.routes"/>
+    <context:component-scan base-package="org.alfresco.enterprise.repo.messaging.camel.routes"/>
 </beans>
