--- src/main/resources/alfresco/subsystems/Events/default/events.properties
@@ -1,4 +1,0 @@
-# Alfresco Events
-
-# Include events?
-alfresco.events.include=CONTENTPUT, NODEADDED, NODEREMOVED, NODEMOVED, NODERENAMED, NODECHECKOUTCANCELLED, NODECHECKEDOUT, NODECHECKEDIN, AUTHADDEDTOGROUP, AUTHREMOVEDFROMGROUP, GROUPDELETED, INHERITPERMISSIONSDISABLED, INHERITPERMISSIONSENABLED, LOCALPERMISSIONGRANTED, LOCALPERMISSIONREVOKED, RECORDCREATED, RECORDREJECTED, FILECLASSIFIED, FILEUNCLASSIFIED
