+++ src/test/resources/log4j.properties
@@ -261,4 +261,4 @@ log4j.logger.org.keycloak=debug
 
 log4j.logger.org.alfresco.repo.content.transform.TransformerDebug=debug
 
-log4j.logger.org.alfresco.repo.rawevents=debug
+log4j.logger.org.alfresco.repo.messaging=debug
