+++ src/test/resources/test-messaging-context.xml
@@ -11,16 +11,16 @@
     <import resource="classpath:/alfresco/subsystems/Messaging/default/messaging-context.xml" />
     
     <bean id="messagingExceptionProcessor" 
-        class="org.alfresco.messaging.camel.MockExceptionProcessor" />
+        class="org.alfresco.repo.messaging.camel.MockExceptionProcessor" />
         
     <bean id="mockConsumer" 
-        class="org.alfresco.messaging.camel.MockConsumer" />
+        class="org.alfresco.repo.messaging.camel.MockConsumer" />
     
     <bean id="mockExceptionThrowingConsumer" 
-        class="org.alfresco.messaging.camel.MockExceptionThrowingConsumer" />
+        class="org.alfresco.repo.messaging.camel.MockExceptionThrowingConsumer" />
     
     <import resource="testRoutes.xml" />
-    <bean id="testNewRouteLoader" class="org.alfresco.messaging.camel.SpringContextRouteLoader">
+    <bean id="testNewRouteLoader" class="org.alfresco.repo.messaging.camel.SpringContextRouteLoader">
         <property name="camelContextId" value="alfrescoCamelContext" />
         <property name="routeContextId" value="customRoutes" />
     </bean>
